pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chunky"
version = "0.1.0"
description = "CLI tool for building local knowledge bases"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
    # Document parsing
    "pdfplumber>=0.10",
    "python-docx>=1.0",
    "python-pptx>=0.6",
    "chardet>=5.0",
    # LLM
    "openai>=1.0",
    "httpx>=0.25",
    # Topic modeling
    "scikit-learn>=1.3",
    "numpy>=1.24",
    # Vector store
    "pymilvus>=2.3",
    "chromadb>=1.5",
]

[project.optional-dependencies]
embedding = ["sentence-transformers>=2.2"]
bertopic = ["bertopic>=0.16"]
milvus-lite = ["milvus-lite>=2.4"]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
dev = [
    "chunky[test,embedding]",
    "ruff>=0.1",
]

[project.scripts]
chunky = "chunky.cli.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
